+++ app/src/main/java/woowacourse/movie/MovieReservationApplication.kt
@@ -1,6 +1,7 @@
 package woowacourse.movie
 
 import android.app.Application
+import woowacourse.movie.database.SettingPreferencesManager
 
 class MovieReservationApplication : Application() {
     override fun onCreate() {
