+++ app/src/main/java/woowacourse/movie/database/ReservationDbHelper.kt
@@ -1,9 +1,8 @@
-package woowacourse.movie.sql
+package woowacourse.movie.database
 
 import android.content.ContentValues
 import android.content.Context
 import android.database.Cursor
-import android.database.sqlite.SQLiteCursor
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 import domain.*
