+++ app/src/main/java/woowacourse/movie/database/SharedPreferenceManager.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.database
 
 interface SharedPreferenceManager {
     fun changeData()
