+++ app/src/main/java/woowacourse/movie/presenter/ReservationListPresenter.kt
@@ -3,7 +3,7 @@ package woowacourse.movie.presenter
 import woowacourse.movie.contract.ReservationListContract
 import woowacourse.movie.model.ReservationUiModel
 import woowacourse.movie.model.mapper.ReservationMapper.toUi
-import woowacourse.movie.sql.ReservationDbHelperInterface
+import woowacourse.movie.database.ReservationDbHelperInterface
 
 class ReservationListPresenter(
     val view: ReservationListContract.View,
