+++ app/src/main/java/woowacourse/movie/receiver/NotificationHelper.kt
@@ -4,7 +4,7 @@ import android.app.Notification
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
-import woowacourse.movie.SettingPreferencesManager
+import woowacourse.movie.database.SettingPreferencesManager
 
 object NotificationHelper {
     fun notifyNotification(
