+++ app/src/main/java/woowacourse/movie/view/widget/SaveStateSpinner.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.widget
 
 import android.R
 import android.os.Bundle
