+++ app/src/main/java/woowacourse/movie/view/widget/SaveStateTextView.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.widget
 
 import android.os.Bundle
 import android.widget.TextView
