+++ app/src/main/java/woowacourse/movie/view/widget/SeatTable.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.widget
 
 import android.widget.TableLayout
 import android.widget.TableRow
