+++ app/src/main/java/woowacourse/movie/view/widget/TimeSpinner.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.widget
 
 import android.widget.Spinner
 import woowacourse.movie.util.LocalFormattedTime
