+++ app/src/test/java/woowacourse/movie/presenter/ReservationListPresenterTest.kt
@@ -7,7 +7,7 @@ import org.junit.Test
 import woowacourse.movie.contract.ReservationListContract
 import woowacourse.movie.model.ReservationUiModel
 import woowacourse.movie.model.mapper.ReservationMapper.toUi
-import woowacourse.movie.sql.ReservationDbHelperInterface
+import woowacourse.movie.database.ReservationDbHelperInterface
 
 class ReservationListPresenterTest {
     private lateinit var presenter: ReservationListPresenter
